Fix GET HTTP 202 status without Location
When a GET request is performed on an asynchronous operation task
monitor URI, the iDRAC service can correctly return the HTTP '202
Accepted' status code without a 'Location' header. [1] This has been
observed while querying the status of the long-running OEM Server
Configuration Profile (SCP) import and export system configuration
operations. [2]

sushy-oem-idrac has incorrectly required the 'Location' header to
continue to query the status of the operation, and determine when it has
completed and whether it succeeded. This fixes that by reusing the URI
provided in the 'Location' header of the response to a previous request,
instead of failing.

[1] http://redfish.dmtf.org/schemas/DSP0266_1.11.0.html#asynchronous-operations
[2] Pushkala Iyer et al., "Server Cloning with Server Configuration
Profiles", (Dell Inc., version 1.1, July 2016),
https://downloads.dell.com/solutions/dell-management-solution-resources/Server
Cloning with Server Configuration Profiles 1_1 Final.pdf.

Co-Authored-By: Derek Higgins <derekh@redhat.com>
Story: 2007640
Task: 39697
Change-Id: I07eef4bab7285d2c454a6112a43b0725e8ee5a09