Update TOX_CONSTRAINTS_FILE for stable/victoria
Update the URL to the upper-constraints file to point to the redirect
rule on releases.openstack.org so that anyone working on this branch
will switch to the correct upper-constraints list automatically when the
requirements repository branches.

That facilitates tests continuing to use the upper-constraints list on
master until the requirements repository has a stable/victoria branch.

Change-Id: I187ad17abf7d8910f99f89df3d3afe77aee5ff86