Update TOX_CONSTRAINTS_FILE for stable/wallaby
Update the URL to the upper-constraints file to point to the redirect
rule on releases.openstack.org so that anyone working on this branch
will switch to the correct upper-constraints list automatically when the
requirements repository branches.

That facilitates tests continuing to use the upper-constraints list on
master until the requirements repository has a stable/wallaby branch.

Change-Id: I9371d27626febd0d5489f066ba9e199819d781cd