Update TOX_CONSTRAINTS_FILE for stable/xena
Update the URL to the upper-constraints file to point to the redirect
rule on releases.openstack.org so that anyone working on this branch
will switch to the correct upper-constraints list automatically when the
requirements repository branches.

That facilitates tests continuing to use the upper-constraints list on
master until the requirements repository has a stable/xena branch.

Change-Id: I0ec5e2c6647340d076a445d14ec1f0253368d08d