Update TOX_CONSTRAINTS_FILE for stable/yoga
Update the URL to the upper-constraints file to point to the redirect
rule on releases.openstack.org so that anyone working on this branch
will switch to the correct upper-constraints list automatically when the
requirements repository branches.

That facilitates tests continuing to use the upper-constraints list on
master until the requirements repository has a stable/yoga branch.

Change-Id: I1758432cb3001da4cbe60a7ab709d578eb272e4f