Add jobs in Paused state to the list of unfinished jobs
Currently, get_unfinished_jobs() returns jobs in Running and Scheduled
states but not Paused state. This causes problems, because a job in
Paused state will still block more instances of the same job from being
created. This change adds jobs in Paused state to this function's
output. It also migrates states indicating that job hasn't finished
to a constant.

Change-Id: I617b7427704aba7bc6cbaa4384906598f9d4e408